backend/requirements.txt
django
djangorestframework
pandas
django-cors-headers
djangorestframework-simplejwt
reportlab